+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples/ResponsiveColumnsMaxWidth.jsx
@@ -48,13 +48,13 @@ const rows = [
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Col_1',
+    displayName: 'Icon',
     width: '60px', // there is no widthUnit prop set in the component, so it will default to px
     align: alignTypes.CENTER,
   },
   {
     id: 'Column-1',
-    displayName: 'Col_2',
+    displayName: 'Service name',
     width: '200px', // will be disregarded because of flexGrow prop
     flexGrow: true, // flexGrow prop has to be set here, as with width set, without that prop the column would not grow.
     maximumWidth: '300px',
@@ -62,7 +62,7 @@ const cols = [
   },
   {
     id: 'Column-2',
-    displayName: 'Col_3',
+    displayName: 'Featured',
     width: '60px', // there is no widthUnit prop set in the component, so it will default to px
     align: alignTypes.CENTER,
   },
@@ -75,6 +75,7 @@ const ResponsiveColumnsMaxWidth = () => (
     columns={cols}
     numberOfColumns={2}
     columnMinimumWidth="234px"
+    rowHeaderIndex={1}
   />
 );