+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples/VerticalFlow.jsx
@@ -44,6 +44,38 @@ const rows = [
       { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
     ],
   },
+  {
+    id: 'row_6',
+    cells: [
+      { content: <IconImage a11yLabel="Picture" height="1.5em" width="1.5em" /> },
+      { content: 'Arterial Sheath Care (6)' },
+      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
+    ],
+  },
+  {
+    id: 'row_7',
+    cells: [
+      { content: ' ' },
+      { content: 'Sbsq Observation Care/Day High Severity 99226 (7)' },
+      { content: ' ' },
+    ],
+  },
+  {
+    id: 'row_8',
+    cells: [
+      { content: <IconImage a11yLabel="Picture" height="1.5em" width="1.5em" /> },
+      { content: 'Arterial Sheath Care (8)' },
+      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
+    ],
+  },
+  {
+    id: 'row_9',
+    cells: [
+      { content: <IconImage a11yLabel="Picture" height="1.5em" width="1.5em" /> },
+      { content: 'Arterial Sheath Care (9)' },
+      { content: <IconFeaturedOutline a11yLabel="Featured" height="1.5em" width="1.5em" /> },
+    ],
+  },
 ];
 
 const getNumberOfColumns = (activeBreakpoint) => {
@@ -73,19 +105,19 @@ const getRowHeight = (activeBreakpoint) => {
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Col_1',
+    displayName: 'Icon',
     width: '60px',
     align: alignTypes.CENTER,
   },
   {
     id: 'Column-1',
-    displayName: 'Col_2',
+    displayName: 'Service name',
     width: '200px', // will be disregarded bacause of flexGrow prop.
     flexGrow: true, // flexGrow prop has to be set here, as with width set, without that prop the column would not grow.
   },
   {
     id: 'Column-2',
-    displayName: 'Col_3',
+    displayName: 'Featured',
     width: '60px', // there is no widthUnit prop set in the component, so it will default to px
     align: alignTypes.CENTER,
   },
@@ -100,6 +132,7 @@ const VerticalFlow = () => {
       columns={cols}
       numberOfColumns={getNumberOfColumns(activeBreakpoint)}
       rowHeight={getRowHeight(activeBreakpoint)}
+      rowHeaderIndex={1}
     />
   );
 };